Fix progress indicators for release/profile builds of web. (#42857)
